+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.12.52 (21-06-2021)
+* PlantUML diagram generation improvements
+* Now there is possibility to generate ER PlantUML diagram for opened metadata.xml file
+
 ## 0.12.51 (17-06-2021)
 * Add support for spread element
 * Improve dependency detection for PlantUML. Also views and fragments are added to the UML diagram as well